Major changes to fix functionality
Major changes to fix functionality, including: Switch from
implementation of LASSO in Scikit Learn to implementation of BP in a
python-port of SPGL1; Switch from use of 1D DCT to 2D IDCT; Change
definition of bigpsi to match definition in literature; Add
reconstruction of y to check performance; Add median filter; Other
changes. The code produces reconstructed images that are consistent with
original images but artifacting is still present.