Final code for project submission
Add code for testing different L1-minimization solvers; comment out code
for images and information that was not needed; and add code for
displaying quantitative measures of algorithm performance. Code was
actually modified between project submission and this commit, but all
edits were minor enough that there should be no change in functionality.